backend/requirements.txt
fastapi
uvicorn[standard]
manim
google-genai
sqlalchemy
psycopg2-binary
imagekitio